refactor: adjust tsky file names to align with class names
