+++ packages/core/src/bsky/Post.ts
@@ -7,7 +7,7 @@ import type {
   AppBskyFeedSearchPosts,
   AppBskyNS,
 } from '@atproto/api'
-import { Paginator } from '~/tsky/paginate'
+import { Paginator } from '~/tsky/Paginator'
 
 export class Post {
   constructor(private instance: AppBskyNS) {}
